updated permission

+++ app/Http/Controllers/SliderController.php
@@ -42,28 +42,33 @@ class SliderController extends Controller
      */
     public function store(Request $request, UploadHelper $helper)
     {
-        $slider = new Slider();
-        if (isset($request->id))
-            $slider = Slider::where('id', $request->id)->first();
+        $user=Auth::user();
+        if($user->typeofuser=='S'|| $user->typeofuser=='V'||$user->typeofuser=='A')
+        {
 
-        switch ($request->action) {
-            case 'remove':
-                $slider->delete();
-                break;
-
-            default:
-                $slider->url = $helper->store($request->file, 'slider');
-                if (isset($request->shop_id)) {
-                    $slider->shop_id = $request->shop_id;
-                }
-                if (isset($request->location) && in_array($request->location, ["home", "shop"])) {
-                    $slider->location = $request->location;
-                }
-                $slider->save();
-                
-                break;
+            $slider = new Slider();
+            if (isset($request->id))
+                $slider = Slider::where('id', $request->id)->first();
+    
+            switch ($request->action) {
+                case 'remove':
+                    $slider->delete();
+                    break;
+    
+                default:
+                    $slider->url = $helper->store($request->file, 'slider');
+                    if (isset($request->shop_id)) {
+                        $slider->shop_id = $request->shop_id;
+                    }
+                    if (isset($request->location) && in_array($request->location, ["home", "shop"])) {
+                        $slider->location = $request->location;
+                    }
+                    $slider->save();
+                    
+                    break;
+            }
         }
-
+        return response()->json(['error' =>'You don\'t have permission to access this resource'],400);
 
         return $slider;
     }
